Collapse PersistablePaymentMethodOption and PaymentMethodIdentifier into single entity
